requirements.txt
streamlit==1.39.0
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
datetime
